moved auth to auth folder
